feat(test): write/read specimen to/from build dir
This commit changes the write/read path of the write_after_read
test function's output/input file so that the file is written to
and read from the build tree in keeping with the best practice of
never modifying the source tree.  In addition to reducing the
likelihood of clobbering a user file, this change also makes it
easier to remove all build/test files with one `rm -rf build`
command.