test(inference_engine_t): more robust subtraction
With this commit, the inference_engine_t type-bound subtraction
operator asserts conformable networks as a precondition.  This
will provide an informative error message in the event that an
operand has one or more components array(s) that is/are not
conformable with the corresponding component array(s) in the
other operand.