feat(inference_engine_t): invariant consistency
This commit enforces inference_engine_t internal self-consistency
pre- and/or postconditions in every inference_engine_t type-bound
procedure by calling the type-bound assert_consistent() subroutine.
With this commit, it should not be possible for any procedure to
receive or produce an inference_engine_t dummy argument that is not
internally self-consistent.