refac(construct_from_json): extensible metadata
This commit refactors the meta_data_t derived type to have a
component string_t array named "key_value" and sized to match the
new list of keys:

 character(len=*), parameter :: key(*) = &
[character(len=len("usingSkipConnections")) :: "modelName", &
 "modelAuthor", "compilationDate", "activationFunction", &
"usingSkipConnections"]

which is passed to findloc() to compute the subscripts for the
corresponding elements of the key_value array, which replaces
previous components named after the keys. This rewfactoring
facilitates adding new key/value pairs by simply adding elements
to the key array.