fix(example): rm dead code; reduce execution time
In the example/train-xor-gate.f90, this commit increases the width
of the network in example/train-xor-gate.f90 to minimize runtime
by trading increased work per training iteration for a reduction
in the number of iterations required to converge.  This change
reduces the execution time by nearly an order of magnitude.