feat(app): restart training from JSON if present
This commit adds a restart feature to complement the recently
added checkpointing feature. If app/train-cloud-microphysics.f90
finds a file with a name of the form <base-name>_network.json,
a trainable_engine_t object will be constructed from the contents
of that file and training will start from that point.