test(adam): more robust convergence
This commit increases the number of epochs in the Adam optimizer
test so that the test is less sensitive to the specific
pseudo-random number sequence employed in initializing the network.
This enables the NAG 7.1 compiler to pass without requiring the
the Fortran 2018 random_init subroutine, which the recently
released 7.2 compiler supports, enabling the 7.2 release to
generate a repeatable pseudo-random number sequence.