test(ifx):fewer training iterations prevents crash
More training iterations in the trainable_engine_test_m function
xor_gate_with_random_weights() yields a test that is less
sensitive to the specific random number sequence used to
initialize network weights.  However, large number of iterations
cause the test suite to crash when compiled with the Intel ifx
compiler due to an apparent compiler bug.  This commit therefore
set the number of iterations for which the test suite currently
passes with the gfortran, ifx, and NAG compilers.