feat(histogram_t): speedup histogram construction
This commit switches the `cloud-microphysics` `histogram_t`
constructor to using a different algorithm when the number of bins
reaches a threshold of 80, at which value the new algorithm
outperforms the original algorithm.  The original algorithm used
`do concurrent`, but required accessing the entire data once per
bin. The new algorithm is inherently serial, but only accesses the
data set once, keeping a running tally of the bin count.