feat(app):filter training data to max info entropy
This commit adds the `--bins` command-line argument, which sets
the number of phase-space hypercubes across the range of each of
the 5 output variables.  Only one data point per bin retained
for use in training a neural network.  This choice maximizes the
information entropy for a given retained training data set size.