feat(train): spread cost sum across iterations
This commit allocates a pair_cost array for each mini-batch so
that each input/output pair's contribution to the cost function
sum can be stored in a separate element rather than keeping a
sequential running tally as the loop through the mini-batch
progresses.  This change lays the groundwork for making the
loop concurrent, which in turn lays the foundation for using
a Fortran 2023 concurrent reduction.  The calulaiton is currently
redundant with a running sum so that a match between the two can
be verified in an assertion.  The reundancy and the assertion will
be removed in a future commit.